database and view
removed extra database entry, fixed view

+++ ViewModel/QuoteVm.cs
@@ -11,6 +11,6 @@ namespace InsuranceQuote.ViewModel
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
-        public int FinalQuote { get; set; }
+        public decimal FinalQuote { get; set; }
     }
 }
